organize component directory
