+++ app/page.js
@@ -1,11 +1,11 @@
 "use client"
 
 import Footer from "@/components/homepage/footer"
-import GetStartedHero from "@/components/homepage/hero/getstartedhero"
-import HeroOne from "@/components/homepage/hero/heroone"
-import HeroThree from "@/components/homepage/hero/herothree"
-import HeroTwo from "@/components/homepage/hero/herotwo"
-import IntroHero from "@/components/homepage/hero/introhero"
+import GetStartedHero from "@/components/homepage/heros/getstartedhero"
+import HeroOne from "@/components/homepage/heros/heroone"
+import HeroThree from "@/components/homepage/heros/herothree"
+import HeroTwo from "@/components/homepage/heros/herotwo"
+import IntroHero from "@/components/homepage/heros/introhero"
 import TopBar from "@/components/homepage/topbar"
 
 export default function HomePage() {
