+++ components/codespace/editor/editorfunctions.js
