+++ components/common/logo.jsx
