+++ components/common/typewritertext.jsx
